v1.1
Incorporates modularity now! Even if you pick DHCP, which it checks that local variable for. Now there's only one bit in the code that's invoking netsh! Okay, twice, but it's conditional and there's no other great way to toggle DHCP or static I guess. There could be a way, but it'd be pushing it and probably make the code more confusing, which I don't want.